@@ -1,14 +1,15 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net8.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+		<IsPackable>false</IsPackable>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\..\ECO\Core\ECO.EventSourcing\ECO.EventSourcing.csproj" />
-    <ProjectReference Include="..\..\ECO\Core\ECO\ECO.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\..\ECO\Core\ECO.EventSourcing\ECO.EventSourcing.csproj" />
+		<ProjectReference Include="..\..\ECO\Core\ECO\ECO.csproj" />
+	</ItemGroup>
 
 </Project>
