@@ -1,13 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFramework>net7.0</TargetFramework>
-		<IsPackable>false</IsPackable>
+		<TargetFramework>net8.0</TargetFramework>
+		<Nullable>enable</Nullable>
 	</PropertyGroup>
 	<ItemGroup>
 		<ProjectReference Include="..\..\ECO\Core\ECO\ECO.csproj" />
 		<ProjectReference Include="..\ECO.Sample.Infrastructure\ECO.Sample.Infrastructure.csproj" />
 	</ItemGroup>
 	<ItemGroup>
-		<PackageReference Include="Resulz" Version="1.3.0" />
+		<PackageReference Include="Resulz" Version="1.5.3" />
 	</ItemGroup>
 </Project>
