@@ -1,12 +1,10 @@
 ﻿using ECO.Data;
 using Microsoft.Extensions.Logging;
-using System;
 
-namespace ECO.Configuration
+namespace ECO.Configuration;
+
+public class DataContextOptions
 {
-    public class DataContextOptions
-    {
-        public Action<IPersistenceUnitFactory, ILoggerFactory> PersistenceUnitFactoryOptions { get; set; } = (persistenceUnitFactory, loggerFactory) => { };
-        public bool RequireTransaction { get; set; } = false;
-    }
+    public Action<IPersistenceUnitFactory, ILoggerFactory> PersistenceUnitFactoryOptions { get; set; } = (persistenceUnitFactory, loggerFactory) => { };
+    public bool RequireTransaction { get; set; } = false;
 }
