@@ -1,26 +1,23 @@
-﻿using System.Collections.Generic;
+﻿namespace ECO.Configuration;
 
-namespace ECO.Configuration
+/// <summary>
+/// ECO Options configuration.
+/// </summary>
+public class ECOOptions
 {
-    /// <summary>
-    /// ECO Options configuration.
-    /// </summary>
-    public class ECOOptions
-    {
-        public const string ECOConfigurationName = "eco";
-        public PersistenceUnitOptions[] PersistenceUnits { get; set; } = new PersistenceUnitOptions[] { };
-    }
+    public const string ECOConfigurationName = "eco";
+    public PersistenceUnitOptions[] PersistenceUnits { get; set; } = [];
+}
 
-    public class PersistenceUnitOptions
-    {
-        public string Name { get; set; } = string.Empty;
+public class PersistenceUnitOptions
+{
+    public string Name { get; set; } = string.Empty;
 
-        public string Type { get; set; } = string.Empty;
+    public string Type { get; set; } = string.Empty;
 
-        public string[] Listeners { get; set; } = new string[] { };
+    public string[] Listeners { get; set; } = [];
 
-        public string[] Classes { get; set; } = new string[] { };
+    public string[] Classes { get; set; } = [];
 
-        public IDictionary<string, string> Attributes { get; set; } = new Dictionary<string, string>();
-    }
+    public IDictionary<string, string> Attributes { get; set; } = new Dictionary<string, string>();
 }
