@@ -1,21 +1,20 @@
-namespace ECO.Data
+namespace ECO.Data;
+
+/// <summary>
+/// Interface that representa a generic persistence manager for an aggregate root and contains the link to the
+/// related persistence unit
+/// </summary>
+/// <typeparam name="T"></typeparam>
+/// <typeparam name="K"></typeparam>
+public interface IPersistenceManager<T, K>
+    where T : class, IAggregateRoot<K>
 {
+    #region Properties
+
     /// <summary>
-    /// Interface that representa a generic persistence manager for an aggregate root and contains the link to the
-    /// related persistence unit
+    /// Related persistence context
     /// </summary>
-    /// <typeparam name="T"></typeparam>
-    /// <typeparam name="K"></typeparam>
-    public interface IPersistenceManager<T, K>
-        where T : class, IAggregateRoot<K>
-    {
-        #region Properties
+    IPersistenceContext PersistenceContext { get; }
 
-        /// <summary>
-        /// Related persistence context
-        /// </summary>
-        IPersistenceContext PersistenceContext { get; }
-
-        #endregion
-    }
+    #endregion
 }
