@@ -1,24 +1,23 @@
-﻿namespace ECO.Data
+﻿namespace ECO.Data;
+
+/// <summary>
+/// Interface that represent a listener that subscribes the different events of a persistence context lifecycle
+/// </summary>
+public interface IPersistenceUnitListener
 {
+    #region Methods
+
     /// <summary>
-    /// Interface that represent a listener that subscribes the different events of a persistence context lifecycle
+    /// Method that occurs before the creation of the persistence context
     /// </summary>
-    public interface IPersistenceUnitListener
-    {
-        #region Methods
+    /// <param name="unit"></param>
+    void ContextPreCreate(IPersistenceUnit unit);
 
-        /// <summary>
-        /// Method that occurs before the creation of the persistence context
-        /// </summary>
-        /// <param name="unit"></param>
-        void ContextPreCreate(IPersistenceUnit unit);
+    /// <summary>
+    /// Method that occurs after the creation of the persistence context
+    /// </summary>
+    /// <param name="unit"></param>
+    void ContextPostCreate(IPersistenceUnit unit, IPersistenceContext context);
 
-        /// <summary>
-        /// Method that occurs after the creation of the persistence context
-        /// </summary>
-        /// <param name="unit"></param>
-        void ContextPostCreate(IPersistenceUnit unit, IPersistenceContext context);
-
-        #endregion
-    }
+    #endregion
 }
