@@ -1,93 +1,87 @@
 using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
 
-namespace ECO.Data
+namespace ECO.Data;
+
+public sealed class PersistenceUnitFactory : IPersistenceUnitFactory
 {
-    public sealed class PersistenceUnitFactory : IPersistenceUnitFactory
+    #region Private_Fields
+
+    private readonly Dictionary<string, IPersistenceUnit> _Units = [];
+
+    private readonly Dictionary<Type, IPersistenceUnit> _Classes = [];
+
+    private readonly ILoggerFactory? _LoggerFactory;
+
+    private readonly ILogger<PersistenceUnitFactory>? _Logger;
+
+    #endregion
+
+    #region ~Ctor
+
+    public PersistenceUnitFactory(ILoggerFactory? loggerFactory = null)
     {
-        #region Private_Fields
+        _LoggerFactory = loggerFactory;
+        _Logger = _LoggerFactory?.CreateLogger<PersistenceUnitFactory>();
+    }
 
-        private readonly IDictionary<string, IPersistenceUnit> _Units = new Dictionary<string, IPersistenceUnit>();
+    #endregion
 
-        private readonly IDictionary<Type, IPersistenceUnit> _Classes = new Dictionary<Type, IPersistenceUnit>();
+    #region Public_Methods
 
-        private readonly ILoggerFactory? _LoggerFactory;
-
-        private readonly ILogger<PersistenceUnitFactory>? _Logger;
-
-        #endregion
-
-        #region ~Ctor
-
-        public PersistenceUnitFactory(ILoggerFactory? loggerFactory = null)
+    public IPersistenceUnitFactory AddPersistenceUnit(IPersistenceUnit persistenceUnit)
+    {
+        ArgumentNullException.ThrowIfNull(persistenceUnit);
+        _Logger?.LogDebug($"Adding persistence unit {persistenceUnit.Name} of type {persistenceUnit.GetType().Name}");
+        _Units.Add(persistenceUnit.Name, persistenceUnit);
+        foreach (var classType in persistenceUnit.Classes)
         {
-            _LoggerFactory = loggerFactory;
-            _Logger = _LoggerFactory?.CreateLogger<PersistenceUnitFactory>();
+            _Classes.Add(classType, persistenceUnit);
         }
+        return this;
+    }
 
-        #endregion
-
-        #region Public_Methods
-
-        public IPersistenceUnitFactory AddPersistenceUnit(IPersistenceUnit persistenceUnit)
+    public IPersistenceUnit GetPersistenceUnit(string name)
+    {
+        if (string.IsNullOrEmpty(name)) throw new ArgumentNullException(nameof(name));
+        if (_Units.TryGetValue(name, out IPersistenceUnit? value))
         {
-            _Logger?.LogDebug($"Adding persistence unit {persistenceUnit.Name} of type {persistenceUnit.GetType().Name}");
-            if (persistenceUnit == null) throw new ArgumentNullException(nameof(persistenceUnit));
-            _Units.Add(persistenceUnit.Name, persistenceUnit);
-            foreach (var classType in persistenceUnit.Classes)
-            {
-                _Classes.Add(classType, persistenceUnit);
-            }
-            return this;
+            return value;
         }
-
-        public IPersistenceUnit GetPersistenceUnit(string name)
+        else
         {
-            if (string.IsNullOrEmpty(name)) throw new ArgumentNullException(nameof(name));
-            if (_Units.ContainsKey(name))
-            {
-                return _Units[name];
-            }
-            else
-            {
-                throw new PersistenceUnitNotFoundException(name);
-            }
+            throw new PersistenceUnitNotFoundException(name);
         }
+    }
 
-        public IPersistenceUnit GetPersistenceUnit<T>()
+    public IPersistenceUnit GetPersistenceUnit<T>()
+    {
+        return GetPersistenceUnit(typeof(T));
+    }
+
+    public IPersistenceUnit GetPersistenceUnit(Type entityType)
+    {
+        ArgumentNullException.ThrowIfNull(entityType);
+        if (_Classes.TryGetValue(entityType, out IPersistenceUnit? value))
         {
-            return GetPersistenceUnit(typeof(T));
+            return value;
         }
-
-        public IPersistenceUnit GetPersistenceUnit(Type entityType)
+        else
         {
-            if (entityType == null) throw new ArgumentNullException(nameof(entityType));
-            if (_Classes.ContainsKey(entityType))
+            foreach (Type registeredType in _Classes.Keys)
             {
-                return _Classes[entityType];
-            }
-            else
-            {
-                foreach (Type registeredType in _Classes.Keys)
+                if (entityType.IsSubclassOf(registeredType))
                 {
-                    if (entityType.IsSubclassOf(registeredType))
-                    {
-                        IPersistenceUnit persistenceUnit = GetPersistenceUnit(registeredType);
-                        _Logger?.LogDebug($"Mapping subclass {entityType.Name} of {registeredType.Name} in persistence unit {persistenceUnit.Name}");
-                        _Classes.Add(entityType, persistenceUnit);
-                        return persistenceUnit;
-                    }
+                    IPersistenceUnit persistenceUnit = GetPersistenceUnit(registeredType);
+                    _Logger?.LogDebug($"Mapping subclass {entityType.Name} of {registeredType.Name} in persistence unit {persistenceUnit.Name}");
+                    _Classes.Add(entityType, persistenceUnit);
+                    return persistenceUnit;
                 }
             }
-            throw new PersistentClassNotRegisteredException(entityType);
         }
-
-        public IDataContext OpenDataContext()
-        {
-            return new DataContext(this, _LoggerFactory?.CreateLogger<DataContext>());
-        }
-
-        #endregion
+        throw new PersistentClassNotRegisteredException(entityType);
     }
+
+    public IDataContext OpenDataContext() => new DataContext(this, _LoggerFactory?.CreateLogger<DataContext>());
+
+    #endregion
 }
