@@ -1,10 +1,9 @@
-﻿namespace ECO
-{
-    /// <summary>
-    /// Interface that defines an entity who is the root of its aggregate
-    /// </summary>
-    public interface IAggregateRoot<T> : IEntity<T>
-    {
+﻿namespace ECO;
 
-    }
+/// <summary>
+/// Interface that defines an entity who is the root of its aggregate
+/// </summary>
+public interface IAggregateRoot<T> : IEntity<T>
+{
+
 }
