@@ -1,33 +1,30 @@
 using ECO.Data;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace ECO
+namespace ECO;
+
+/// <summary>
+/// Interface that defines a read only repository
+/// </summary>
+/// <typeparam name="T"></typeparam>
+/// <typeparam name="K"></typeparam>
+public interface IReadOnlyRepository<T, K> : IQueryable<T>, IPersistenceManager<T, K>
+    where T : class, IAggregateRoot<K>
 {
+    #region
+
     /// <summary>
-    /// Interface that defines a read only repository
+    /// Method that loads the entity from the repository
     /// </summary>
-    /// <typeparam name="T"></typeparam>
-    /// <typeparam name="K"></typeparam>
-    public interface IReadOnlyRepository<T, K> : IQueryable<T>, IPersistenceManager<T, K>
-        where T : class, IAggregateRoot<K>
-    {
-        #region
+    /// <param name="identity"></param>
+    /// <returns></returns>
+    T? Load(K identity);
 
-        /// <summary>
-        /// Method that loads the entity from the repository
-        /// </summary>
-        /// <param name="identity"></param>
-        /// <returns></returns>
-        T? Load(K identity);
+    /// <summary>
+    /// Method that loads asynchronously the entity from the repository
+    /// </summary>
+    /// <param name="identity"></param>
+    /// <returns></returns>
+    Task<T?> LoadAsync(K identity);
 
-        /// <summary>
-        /// Method that loads asynchronously the entity from the repository
-        /// </summary>
-        /// <param name="identity"></param>
-        /// <returns></returns>
-        Task<T?> LoadAsync(K identity);
-
-        #endregion
-    }
+    #endregion
 }
