@@ -1,18 +1,17 @@
-﻿namespace ECO
+﻿namespace ECO;
+
+/// <summary>
+/// Interface that defines a versionble Aggregate root
+/// </summary>
+/// <typeparam name="T"></typeparam>
+public interface IVersionableAggregateRoot<T> : IAggregateRoot<T>
 {
+    #region Properties
+
     /// <summary>
-    /// Interface that defines a versionble Aggregate root
+    /// Version of the aggregate root
     /// </summary>
-    /// <typeparam name="T"></typeparam>
-    public interface IVersionableAggregateRoot<T> : IAggregateRoot<T>
-    {
-        #region Properties
+    int Version { get; }
 
-        /// <summary>
-        /// Version of the aggregate root
-        /// </summary>
-        int Version { get; }
-
-        #endregion
-    }
+    #endregion
 }
