@@ -1,18 +1,17 @@
-namespace ECO
+namespace ECO;
+
+/// <summary>
+/// Interface that defines a versionble Entity
+/// </summary>
+/// <typeparam name="T"></typeparam>
+public interface IVersionableEntity<T> : IEntity<T>
 {
+    #region Properties
+
     /// <summary>
-    /// Interface that defines a versionble Entity
+    /// Version of the entity
     /// </summary>
-    /// <typeparam name="T"></typeparam>
-    public interface IVersionableEntity<T> : IEntity<T>
-    {
-        #region Properties
+    int Version { get; }
 
-        /// <summary>
-        /// Version of the entity
-        /// </summary>
-        int Version { get; }
-
-        #endregion
-    }
+    #endregion
 }
