@@ -1,13 +1,10 @@
-﻿using System;
+﻿namespace ECO.Data;
 
-namespace ECO.Data
+public interface IPersistenceUnitFactory
 {
-    public interface IPersistenceUnitFactory
-    {
-        IPersistenceUnitFactory AddPersistenceUnit(IPersistenceUnit persistenceUnit);
-        IPersistenceUnit GetPersistenceUnit(string name);
-        IPersistenceUnit GetPersistenceUnit(Type entityType);
-        IPersistenceUnit GetPersistenceUnit<T>();
-        IDataContext OpenDataContext();
-    }
+    IPersistenceUnitFactory AddPersistenceUnit(IPersistenceUnit persistenceUnit);
+    IPersistenceUnit GetPersistenceUnit(string name);
+    IPersistenceUnit GetPersistenceUnit(Type entityType);
+    IPersistenceUnit GetPersistenceUnit<T>();
+    IDataContext OpenDataContext();
 }
