@@ -1,20 +1,8 @@
-﻿using System;
+﻿namespace ECO.Data;
 
-namespace ECO.Data
+public class PersistenceUnitNotFoundException(string persistenceUnitName) : ApplicationException($"Persistence Unit '{persistenceUnitName}' not found")
 {
-    public class PersistenceUnitNotFoundException : ApplicationException
-    {
-        #region Public_Properties
-
-        public string PersistenceUnitName { get; protected set; }
-
-        #endregion
-
-        #region Ctor
-
-        public PersistenceUnitNotFoundException(string persistenceUnitName)
-            : base($"Persistence Unit '{persistenceUnitName}' not found") => PersistenceUnitName = persistenceUnitName;
-
-        #endregion
-    }
+    #region Public_Properties
+    public string PersistenceUnitName { get; protected set; } = persistenceUnitName;
+    #endregion
 }
