@@ -1,20 +1,17 @@
-using System;
+namespace ECO;
 
-namespace ECO
+/// <summary>
+/// Interface that defines an Entity with an identifier of T type
+/// </summary>
+/// <typeparam name="T"></typeparam>
+public interface IEntity<T> : IEquatable<IEntity<T>>
 {
+    #region Properties
+
     /// <summary>
-    /// Interface that defines an Entity with an identifier of T type
+    /// Identifier of the entity
     /// </summary>
-    /// <typeparam name="T"></typeparam>
-    public interface IEntity<T> : IEquatable<IEntity<T>>
-    {
-        #region Properties
+    T? Identity { get; }
 
-        /// <summary>
-        /// Identifier of the entity
-        /// </summary>
-        T Identity { get; }
-
-        #endregion
-    }
+    #endregion
 }
