@@ -1,12 +1,9 @@
-﻿using System;
+﻿namespace ECO;
 
-namespace ECO
+/// <summary>
+/// Interface that defines a value object
+/// </summary>    
+public interface IValueObject<T> : IEquatable<T>
 {
-    /// <summary>
-    /// Interface that defines a value object
-    /// </summary>    
-    public interface IValueObject<T> : IEquatable<T>
-    {
 
-    }
 }
